Fix selectedValue for NavDrawer [packages-only]
