+++ .github/workflows/ci-releases.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - releases/1.10
-      - feat/ideamodule
     paths:
       - 'SharePointFramework/**'
       - 'Install/**'
