+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/IdeaModule.module.scss
@@ -6,7 +6,7 @@
 
   .content {
     flex: 1;
-    padding: 32px;
+    padding: 16px 32px;
     gap: 16px;
     display: grid;
     justify-content: flex-start;
