+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/IdeaModule.tsx
@@ -3,14 +3,46 @@ import styles from './IdeaModule.module.scss'
 import { IdeaModuleContext } from './context'
 import { IIdeaModuleProps } from './types'
 import { useIdeaModule } from './useIdeaModule'
-import { Divider, FluentProvider, IdPrefixProvider, Spinner } from '@fluentui/react-components'
-import { customLightTheme, UserMessage } from 'pp365-shared-library'
+import {
+  Divider,
+  FluentProvider,
+  IdPrefixProvider,
+  Spinner,
+  Tooltip
+} from '@fluentui/react-components'
+import { customLightTheme, getFluentIcon, setUrlHash, UserMessage } from 'pp365-shared-library'
+import {
+  NavCategory,
+  NavCategoryItem,
+  NavDrawer,
+  NavDrawerBody,
+  NavDrawerHeader,
+  NavItem,
+  NavSectionHeader,
+  NavSubItem,
+  NavSubItemGroup,
+  NavDivider,
+  AppItemStatic
+} from '@fluentui/react-nav-preview'
+import {
+  Board20Filled,
+  Board20Regular,
+  NotePin20Filled,
+  NotePin20Regular,
+  Lightbulb20Filled,
+  Lightbulb20Regular,
+  bundleIcon
+} from '@fluentui/react-icons'
 import { IdeaField } from './IdeaField'
 import { IdeaPhaseBar } from './IdeaPhaseBar'
-import { IdeaNav } from './IdeaNav'
+
+const Dashboard = bundleIcon(Board20Filled, Board20Regular)
+const Lightbulb = bundleIcon(Lightbulb20Filled, Lightbulb20Regular)
+const JobPostings = bundleIcon(NotePin20Filled, NotePin20Regular)
 
 export const IdeaModule: FC<IIdeaModuleProps> = (props) => {
-  const { state, setState, isOpen, renderHamburger, fluentProviderId } = useIdeaModule(props)
+  const { state, setState, getSelectedIdea, isOpen, renderHamburger, fluentProviderId } =
+    useIdeaModule(props)
 
   return (
     <IdeaModuleContext.Provider value={{ props, state, setState }}>
@@ -20,7 +52,73 @@ export const IdeaModule: FC<IIdeaModuleProps> = (props) => {
             <Spinner label='Laster inn idémodulen' size='extra-large' />
           ) : (
             <div className={styles.ideaModule}>
-              <IdeaNav />
+              <NavDrawer
+                  selectedValue={`nav${state.selectedIdea?.item.Id.toString()}`}
+                selectedCategoryValue={state.selectedIdea?.item.processing ? 'behandlingIdeer' : 'registreringIdeer'}
+                openCategories={['registreringIdeer', 'behandlingIdeer']}
+                open={isOpen}
+                type='inline'
+                size='small'
+              >
+                <NavDrawerHeader>
+                  <Tooltip content='Navigation' relationship='label'>
+                    {renderHamburger()}
+                  </Tooltip>
+                </NavDrawerHeader>
+                <NavDrawerBody>
+                  <AppItemStatic icon={getFluentIcon('Lightbulb')}>Idémodul</AppItemStatic>
+                  <NavItem href='#' icon={<Dashboard />} value='total'>
+                    Totaloversikt
+                  </NavItem>
+                  <NavSectionHeader>Registrering</NavSectionHeader>
+                  <NavItem href='#' icon={<Dashboard />} value='registrering'>
+                    Oversikt
+                  </NavItem>
+                  <NavCategory value='registreringIdeer'>
+                    <NavCategoryItem icon={<Lightbulb />}>Mine idéer</NavCategoryItem>
+                    <NavSubItemGroup>
+                      {state.ideas.data.items
+                        .filter((idea) => !idea.processing)
+                        .map((idea) => (
+                          <NavSubItem
+                            key={idea.Id.toString()}
+                            value={`nav${idea.Id.toString()}`}
+                            onClick={() => {
+                              setUrlHash({ ideaId: idea.Id.toString() })
+                              getSelectedIdea()
+                            }}
+                          >
+                            {idea?.Title}
+                          </NavSubItem>
+                        ))}
+                    </NavSubItemGroup>
+                  </NavCategory>
+                  <NavDivider />
+                  <NavSectionHeader>Behandling</NavSectionHeader>
+                  <NavItem href='#' icon={<Dashboard />} value='behandling'>
+                    Oversikt
+                  </NavItem>
+                  <NavCategory value='behandlingIdeer'>
+                    <NavCategoryItem icon={<JobPostings />}>Mine idéer</NavCategoryItem>
+                    <NavSubItemGroup>
+                      {state.ideas.data.items
+                        .filter((idea) => idea.processing)
+                        .map((idea) => (
+                          <NavSubItem
+                            key={idea.Id.toString()}
+                            value={`nav${idea.Id.toString()}`}
+                            onClick={() => {
+                              setUrlHash({ ideaId: idea.Id.toString() })
+                              getSelectedIdea()
+                            }}
+                          >
+                            {idea?.Title}
+                          </NavSubItem>
+                        ))}
+                    </NavSubItemGroup>
+                  </NavCategory>
+                </NavDrawerBody>
+              </NavDrawer>
               <div className={styles.content}>
                 {state.error && (
                   <UserMessage
