--- SharePointFramework/PortfolioWebParts/src/components/IdeaModule/IdeaNav/index.ts
@@ -1,1 +1,0 @@
-export * from './IdeaNav'
