+++ SharePointFramework/PortfolioWebParts/src/components/IdeaModule/useIdeaModule.tsx
@@ -13,7 +13,7 @@ import { Hamburger } from '@fluentui/react-nav-preview'
  * Component logic hook for `IdeaModule` component.
  *
 clear */
-export function useIdeaModule(props?: IIdeaModuleProps) {
+export function useIdeaModule(props: IIdeaModuleProps) {
   const { state, setState } = useIdeaModuleState(props)
   const fluentProviderId = useId('fp-idea-module')
